+++ app/src/main/java/com/example/leaveit/presentation/placeview/showplaceview/ShowPlaceModel.kt
@@ -1,4 +1,4 @@
-package com.example.leaveit.presentation.placeview
+package com.example.leaveit.presentation.placeview.showplaceview
 
 data class ShowPlaceModel(
     val contentId : Int,
