+++ app/src/main/java/com/example/leaveit/presentation/placeview/showplaceview/ShowPlaceModelMapper.kt
@@ -1,4 +1,4 @@
-package com.example.leaveit.presentation.placeview
+package com.example.leaveit.presentation.placeview.showplaceview
 
 interface ShowPlaceModelMapper<ShowPlaceModel> {
     fun toPlaceModel() : ShowPlaceModel
