+++ app/src/main/java/com/example/leaveit/presentation/placeview/showplaceview/ShowPlaceView.kt
@@ -1,20 +1,15 @@
-package com.example.leaveit.presentation.placeview
+package com.example.leaveit.presentation.placeview.showplaceview
 
 import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.FrameLayout
-import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
-import androidx.lifecycle.ViewModel
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.leaveit.databinding.FragmentShowplaceviewBinding
-import com.example.leaveit.domain.AppModule
 import com.example.leaveit.domain.testHiltClass
-import com.example.leaveit.remote.api.RetrofitService
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
